Create HackRFOne WiFi
This code initializes the HackRF, sets the frequency range to 2.4-2.5 GHz, the sample rate to 2 MHz, and the bandwidth to 2 MHz. It also sets the LNA gain to 16 dB and the VGA gain to 24 dB. Then it starts receiving data and prints it to the console.

To view nearby WiFi networks, you will need to use a software tool to analyze the data that is being received by the HackRF. Some popular options for this are Wireshark and Kismet.